+++ game_patch/in_game_ui/spectate_mode.h
@@ -9,6 +9,8 @@ namespace rf
 void SpectateModeSetTargetPlayer(rf::Player* player);
 void SpectateModeInit();
 void SpectateModeAfterFullGameInit();
+void SpectateModeLevelInit();
 void SpectateModeDrawUI();
 void SpectateModeOnDestroyPlayer(rf::Player* player);
+void SpectateModePlayerCreateEntityPost(rf::Player* player);
 bool SpectateModeIsActive();